WiFi: "Connect" option disappear after rotating.
The connect option disappears because there is a problem in saving and
restoring the state of the access point showed in the dialog box. When a
rotation accurs, the framework calls onSaveInstanceState(Bundle outState)
of the wifisettings activity. In this method, the state of the access is
saved, but not its rssi level. After a rotation, when restoring the access
point state, the rssi is initialized to the default value, which means out of
range. This is why the connect option is not displayed. I fixed the issue by
saving and restoring the rssi level of the access point.

Change-Id: Ibd58132c922828941603673383ff81f25d61556e
Orig-Change-Id: I624cc8ce82a69fd94931088dc0f7833663ef0772
Signed-off-by: Paul Naoumenko <paulx.naoumenko@intel.com>
Signed-off-by: Jean-Marie Lemetayer <jean-mariex.lemetayer@intel.com>
Signed-off-by: Jian Luo <jian.luo@intel.com>
Signed-off-by: Bruce Beare <bruce.j.beare@intel.com>
Signed-off-by: Jack Ren <jack.ren@intel.com>
Author-tracking-BZ: 34109