Settings: fcharge for all (2/2)
Thanks roman for the basic idea to use an overlay
to set the fast charge path so each device can set
the propper path to the custom kernel.
As well this commit pays attention if the user
flashed a kernel which does not support fastcharge
or is using another path. Option for tile in settings will be removed
as well.